more unit tests (and activated some skipped tests)

+++ src/test/java/nom/tam/util/test/ArrayFuncs2Test.java
@@ -581,7 +581,7 @@ public class ArrayFuncs2Test {
 
     @Test
     public void copyDoubleInto() {
-        float[][] deepArray = new float[][]{
+        double[][] deepArray = new double[][]{
             {
                 1,
                 2,
